Move validation error helpers out of ErrorFactories
changelog_begin
changelog_end
