+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/ledger/api/validation/ValueValidator.scala
@@ -10,13 +10,13 @@ import com.daml.ledger.api.v1.{value => api}
 import com.daml.lf.data._
 import com.daml.lf.value.Value.{ContractId, ValueUnit}
 import com.daml.lf.value.{Value => Lf}
-import com.daml.platform.server.api.validation.{ErrorFactories, FieldValidations}
+import com.daml.platform.server.api.validation.FieldValidations
 import io.grpc.StatusRuntimeException
 import scalaz.std.either._
 import scalaz.syntax.bifunctor._
 
 object ValueValidator {
-  import ErrorFactories._
+  import ValidationErrors._
   import FieldValidations._
 
   private[validation] def validateRecordFields(
