+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/api/services/grpc/GrpcHealthService.scala
@@ -13,7 +13,6 @@ import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.platform.api.grpc.GrpcApiService
 import com.daml.platform.server.api.DropRepeated
 import com.daml.platform.server.api.services.grpc.GrpcHealthService._
-import com.daml.platform.server.api.validation.ErrorFactories
 import io.grpc.ServerServiceDefinition
 import io.grpc.health.v1.health.{
   HealthAkkaGrpc,
@@ -41,7 +40,7 @@ class GrpcHealthService(
   private val errorLogger: ContextualizedErrorLogger =
     new DamlContextualizedErrorLogger(logger, loggingContext, None)
 
-  import ErrorFactories.invalidArgument
+  import com.daml.ledger.api.validation.ValidationErrors.invalidArgument
 
   override def bindService(): ServerServiceDefinition =
     HealthGrpc.bindService(this, executionContext)
