+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/api/validation/FieldValidations.scala
@@ -13,9 +13,8 @@ import com.daml.lf.data.Ref.Party
 import com.daml.lf.value.Value.ContractId
 import io.grpc.StatusRuntimeException
 
-// TODO error codes: Remove default usage of ErrorFactories
 object FieldValidations {
-  import ErrorFactories._
+  import com.daml.ledger.api.validation.ValidationErrors._
 
   def matchLedgerId(
       ledgerId: LedgerId
